Big update: Split the pipeline into functions with specific outputs.
By creating a Benchmark class and checking for specific filenames
within a new project folder, the program can assess the progress of
our work in that directory. It then offers the user a choice of which
benchmark they want to start up from. Still unfinished, but the
seperate functions are on their way and being test. Several tasks have
been simplified and streamlined.

The setup file was also updated to account for packages no longer used.