combine multiple steps with new solution to cluster points.
I woke up with this idea all of a sudden: in addition to using the
fly_eye find ommatidia function for finding the center of ommatidias
in a cross-section or shell of the data, we can then find all the
coordinates corresponding to that center using their distance. We can
iterate this over multiple layers or shells to find the whole cluster,
or tips of it, overcoming the effect of overlap. We should consider doing
>2 shells in the case of very non-spherical eyes, where the lenses may
overlap even when only considering half of the shell. The number of
shells is limited by the number of data points avaiable from each
lens, which is limited by the resolution of the scan and the number of ommatidia.