Improve IO calculations to account for previously noisy results.
Previous IO calculations used the algorithm for finding the angle
between skew vectors in 3D but this allowed for approximate
intersections (points of minimal distance) that could vary
substantially. Instead, based on Wardill 2017 (the robber fly paper),
for the anatomical IO angle, we first find a plane using svd on the
union of the two clusters. Then, we project the points onto that plane
and find the intersection of the principle components of each cluster
on this 2d plane. Hopefully, this will restrict the problem enough to
minimize our variable results.