Reverted back to old method of inside and outside shells.
The new method works alright for small datasets but for the eyes with
many (> 10000) facets, it is impossibly slow. Also, though the new
method segmented the bee cones sucessfully, the IO angles were way too
variable to even be useful. So we are marketing this method primarily
for spherical eyes, though the new method can work for measuring the
skewness of the ommatidia - allowing the measurement of effective
ommatidial diameter.